pyproject.toml
[tool.poetry]
name = "legal-contract-extractor"
version = "1.0.0"
description = "Production-grade NLP system for extracting structured entities from legal PDF contracts"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

# ── Web Framework ─────────────────────────────────────────────────────────────
# FastAPI 0.111+ requires fastapi-cli which requires typer >=0.12
# spaCy 3.8+ also dropped the typer <0.10 cap — these are now compatible
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-multipart = "^0.0.9"

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
PyYAML = "^6.0.1"
python-json-logger = "^2.0.7"

# ── PDF Processing ────────────────────────────────────────────────────────────
pdfplumber = "^0.11.0"
pdf2image = "^1.17.0"
pytesseract = "^0.3.10"
Pillow = "^10.3.0"
opencv-python-headless = "^4.9.0"

# ── NLP ───────────────────────────────────────────────────────────────────────
# spaCy 3.8 — dropped typer <0.10 upper bound, now compatible with fastapi-cli
spacy = "^3.8.0"

# transformers 4.44+ requires torch >=2.4 — keep these in sync
# tokenizers must stay within what transformers declares (auto-resolved by Poetry)
transformers = "^4.44.0"
tokenizers = "^0.19.0"
datasets = "^2.19.0"
evaluate = "^0.4.2"
seqeval = "^1.2.2"

# ── ML / Experiment Tracking ──────────────────────────────────────────────────
# mlflow 2.12+ has full pydantic v2 support
mlflow = "^2.12.1"

# ── PyTorch CPU ───────────────────────────────────────────────────────────────
# torch 2.4+ required by transformers >=4.44
# Sourced from PyTorch's own index to get the correct CPU-only wheel
torch = {version = "^2.4.0", source = "pytorch-cpu"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"


[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "WARNING"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"